create a cpp file for HDF5FileLayout
